+++ module/actor-sheet.js
@@ -152,7 +152,7 @@ export class GurpsActorSheet extends ActorSheet {
             const w = 50
             const h = 50
             const preview = DragDrop.createDragImage(img, w, h)
-            ev.dataTransfer.setDragImage(preview, w / 2, h / 2)
+            ev.dataTransfer.setDragImage(preview, 0, 0)
           }
           let newd = {
             actorid: this.actor.id,
@@ -492,20 +492,17 @@ export class GurpsActorSheet extends ActorSheet {
       ev.preventDefault()
       let parent = $(ev.currentTarget).closest('[data-key]')
       let path = parent.attr('data-key')
-
-      let eqt = duplicate(getProperty(this.actor.data, path))
+      let eqt = getProperty(this.actor.data, path)
       let value = parseInt(eqt.count) + (ev.shiftKey ? 5 : 1)
       if (isNaN(value)) value = 0
-      eqt.count = value
-      await this.actor.update({ [path]: eqt })
-      await this.actor.updateParentOf(path)
+      await this.actor.updateEqtCount(path, value)
     })
     html.find('button[data-operation="equipment-dec"]').click(async ev => {
       ev.preventDefault()
       let parent = $(ev.currentTarget).closest('[data-key]')
       let path = parent.attr('data-key')
       let actor = this.actor
-      let eqt = duplicate(getProperty(actor.data, path))
+      let eqt = getProperty(actor.data, path)
       if (eqt.count == 0) {
         let agree = false
         await Dialog.confirm({
@@ -517,9 +514,7 @@ export class GurpsActorSheet extends ActorSheet {
       } else {
         let value = parseInt(eqt.count) - (ev.shiftKey ? 5 : 1)
         if (isNaN(value) || value < 0) value = 0
-        eqt.count = value
-        await this.actor.update({ [path]: eqt })
-        await this.actor.updateParentOf(path)
+        await this.actor.updateEqtCount(path, value)
       }
     })
 
@@ -901,7 +896,7 @@ export class GurpsActorSheet extends ActorSheet {
       let targetkey = element.dataset.key
       if (!!targetkey) {
         let srckey = dragData.key
-        this.actor.moveEquipment(srckey, targetkey)
+        this.actor.moveEquipment(srckey, targetkey, event.shiftKey)
       }
     }
   }
@@ -1510,17 +1505,17 @@ export class GurpsActorNpcSheet extends GurpsActorSheet {
   }
 }
 
-let _getProperty = function (object, key) {
-  let target = object
-
-  // Convert the key to an object reference if it contains dot notation
-  if (key.indexOf('.') !== -1) {
-    let parts = key.split('.')
-    key = parts.pop()
-    target = parts.reduce((o, i) => {
-      if (!o.hasOwnProperty(i)) o[i] = {}
-      return o[i]
-    }, object)
+export class GurpsInventorySheet extends GurpsActorSheet {
+  /** @override */
+  static get defaultOptions() {
+    return mergeObject(super.defaultOptions, {
+      classes: ['gurps', 'sheet', 'actor'],
+      template: 'systems/gurps/templates/inventory-sheet.html',
+      width: 700,
+      height: 400,
+      tabs: [],
+      scrollY: [],
+      dragDrop: [{ dragSelector: 'item-list .item', dropSelector: null }],
+    })
   }
-  return target
 }